### 2022-07-18 #4590 @editorialbot
[REVIEW]: MARTINI: The Little Match and Replace Tool for Automatic Code Rewriting
**Submitting author:** @ajohnson-uoregon (<a href="http://orcid.org/0000-0002-7199-8221">Alister Johnson</a>)
**Repository:** https://github.com/ajohnson-uoregon/clang-rewrite-only/
**Branch with paper.md** (empty if default branch): joss22
**Version:** joss22+europar22
**Archive:** https://doi.org/10.6084/m9.figshare.20483967.v1
