### 2022-07-19 #4590 @ajohnson-uoregon
Per the suggestion in the pre-review thread, I updated the bibliography to only have things that are actually cited so editorialbot doesn't go looking for dois.
