### 2022-07-20 ajohnson-uoregon/clang-rewrite-only#1 @wimvanderbauwhede
JOSS#4590 Link or integrate Instructions.pdf into README
[openjournals/joss-reviews#4590](https://github.com/openjournals/joss-reviews/issues/4590)

The Instructions.pdf is very clear, please link to it from the README, or if possible integrate it into the README.
