### 2022-07-20 ajohnson-uoregon/clang-rewrite-only#2 @wimvanderbauwhede
JOSS#4590 no plain-text LICENSE file
[openjournals/joss-reviews#4590](https://github.com/openjournals/joss-reviews/issues/4590)

There is no plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license at top level or in the clang-rewrite-only folder, I checked the master and joss22 branches. This is a JOSS requirement so please add a  LICENSE file.
