### 2022-07-20 #4590 @wimvanderbauwhede
@danielskatz @ajohnson-uoregon If I read it correctly, according to [https://llvm.org/docs/DeveloperPolicy.html#id34](https://llvm.org/docs/DeveloperPolicy.html#id34), developers can still choose the [OSI approved "The University of Illinois/NCSA Open Source License (NCSA)"](https://opensource.org/licenses/UoI-NCSA.php). If that is acceptable, I'd propose to add this license to the sources in `clang-rewrite`.
