### 2022-07-20 ajohnson-uoregon/clang-rewrite-only#2 @ajohnson-uoregon
A license file and license headers in all the code files have been added: https://github.com/ajohnson-uoregon/clang-rewrite-only/blob/joss22/clang-tools-extra/clang-rewrite/LICENSE.txt

LLVM as a whole is licensed under [Apache 2.0 with LLVM Exceptions](https://llvm.org/LICENSE.txt), so I chose to place this under the same license since it'll eventually (hopefully) get upstreamed into LLVM. I don't see that in the OSI approved list, but since it's essentially a more permissive version of Apache 2.0 is this okay or should I change it to vanilla Apache 2.0?
