### 2022-07-21 ajohnson-uoregon/clang-rewrite-only#2 @wimvanderbauwhede
[openjournals/joss-reviews#4590](https://github.com/openjournals/joss-reviews/issues/4590)

@ajohnson-uoregon
I think the LLVM project requires either the "Apache 2.0 License with LLVM exceptions" or the old NCSA license, I don't think vanilla Apache 2.0 is acceptable.

For JOSS your license has to be OSI approved. Can you maybe use the NCSA license? https://opensource.org/licenses/UoI-NCSA.php
