### 2022-07-23 8cac91a
updating readme with instructions
