+++ clang-tools-extra/clang-rewrite/README.md
@@ -1,5 +1,193 @@
 # clang-rewrite Readme
 
+## How-To
+
+A (slightly outdated) PDF of these instructions is available [here](https://github.com/ajohnson-uoregon/llvm-project/blob/feature-ajohnson/clang-tools-extra/clang-rewrite/Instructions.pdf).
+
+### Getting Started
+
+Required dependencies:
+- [CMake](https://cmake.org/install/) ≥ 3.13.4
+
+Optional dependencies:
+- [Ninja](https://ninja-build.org/) build system
+
+##### Building `clang-rewrite`/MARTINI:
+1. Install CMake and Ninja (recommended for quicker build of LLVM, but not
+required).
+2. Clone our fork of the LLVM monorepo.
+  - `git clone https://github.com/ajohnson-uoregon/llvm-project.git`
+3. Check out `europar22-artifact` branch (for version used in that paper) or
+  `feature-ajohnson` branch (for development version; may not always work).
+  - `git checkout europar22-artifact`
+4. Follow instructions to build LLVM, including `clang` and `clang-tools-extra`
+projects. Full instructions: https://llvm.org/docs/GettingStarted.html. Short
+instructions:
+  - Enter `llvm-project` directory.
+  - `mkdir build`
+  - `cd build`
+  - For Ninja (recommended):
+    -  
+        ```
+        cmake -G Ninja \
+        -DLLVM_TARGETS_TO_BUILD="X86;NVPTX;AMDGPU" \
+        -DCMAKE_INSTALL_PREFIX="<install dir>" \
+        -DCMAKE_BUILD_TYPE=RelWithDebInfo \
+        -DLLVM_ENABLE_PROJECTS="clang;clang-tools-extra" \
+        ../llvm
+        ```
+    - `ninja -j <N> install`
+  - For Unix make:
+    -  
+        ```
+        cmake -G "Unix Makefiles" \
+        -DLLVM_TARGETS_TO_BUILD="X86;NVPTX;AMDGPU" \
+        -DCMAKE_INSTALL_PREFIX="<install dir>" \
+        -DCMAKE_BUILD_TYPE=RelWithDebInfo \
+        -DLLVM_ENABLE_PROJECTS="clang;clang-tools-extra" \
+        ../llvm
+        ```
+    - `cmake --build . -j <N> --target install`
+5. The MARTINI/`clang-rewrite` executable will be in the install directory chosen
+when building LLVM.
+
+#### Notes on Build
+
+- The majority of our contributions are in `llvm-project/clang-tools-extra/clang-rewrite`.
+- LLVM/Clang is a _very_ large project, and will take several hours to build
+even when done in parallel.
+- Parallel builds are _highly_ recommended, but will use a large amount of memory when linking the final libraries and executables. The last approx. 500 targets may need to be built in serial if your machine has ≤24GB of memory.
+
+### Instructions for Running Tests
+
+These tests should reproduce the results from our EuroPar '22 paper. As of July 2022, we recommend running these tests with the version of the code from EuroPar,
+as the `feature-ajohnson` branch is under active development and thus somewhat unstable and may not always work.
+
+#### modernize-use-nullptr
+
+In `llvm-project/clang-tools-extra/clang-rewrite/tests/modernize-use-nullptr`,
+run:
+- `clang-rewrite --inst-file=spec.cpp test.cpp --`
+
+The file `spec.cpp` contains the matchers and replacers used to emulate
+modernize-use-nullptr’s functionality. The file `test.cpp` contains the code
+to be rewritten and the expected result.
+
+The above command should only take a few seconds and produce a file named
+`test.test_final.cpp` with the rewritten code (and expected result for comparison).
+This example corresponds to the one presented in the paper’s Introduction.
+
[…]
