### 2022-07-23 ajohnson-uoregon/clang-rewrite-only#1 @ajohnson-uoregon
Done: 8cac91a34b81e95e2d62e0eac941ac3f9d8ecd0e (this repo) and [0a882a70](https://github.com/ajohnson-uoregon/llvm-project/commit/0a882a703fd58337d3073e026c618cbe0643a922) (main repo).

Markdown formatting is still a little janky but the information is there.
