+++ clang-tools-extra/clang-rewrite/LocalASTMatchers.h
@@ -1,11 +1,3 @@
-//===--- LocalASTMatchers.h - AST matchers that don't warrant ASTMatchers.h ==//
-//
-// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-//===----------------------------------------------------------------------===//
-
 #ifndef CLANG_REWRITE_LOCAL_AST_MATCHERS_H
 #define CLANG_REWRITE_LOCAL_AST_MATCHERS_H
 
@@ -29,7 +21,7 @@ namespace ast_matchers {
 /// \endcode
 /// despite the return value being an IntegerLiteral.
 
-/*AST_MATCHER_P(ReturnStmt, hasExpectedReturnType, internal::Matcher<QualType>,
+AST_MATCHER_P(ReturnStmt, hasExpectedReturnType, internal::Matcher<QualType>,
               InnerMatcher) {
   const auto &Parents = Finder->getASTContext().getParents(Node);
 
@@ -51,7 +43,7 @@ namespace ast_matchers {
   } else {
     return InnerMatcher.matches(FuncDeclNode->getReturnType(), Finder, Builder);
   }
-}*/
+}
 
 /// Matches the first argument (grid dim) in <<<>>> on CUDA kernel calls.
 ///
