+++ clang-tools-extra/clang-rewrite/ConstructMatchers.h
@@ -1,20 +1,19 @@
-//===--- ConstructMatchers.h - Dynamically create AST matchers ------------===//
-//
-// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-//===----------------------------------------------------------------------===//
-
 #ifndef CLANG_CONSTRUCT_MATCHERS_H
 #define CLANG_CONSTRUCT_MATCHERS_H
 
+#include "clang/ASTMatchers/Dynamic/Registry.h"
+#include "clang/ASTMatchers/ASTMatchers.h"
+#include "clang/ASTMatchers/ASTMatchersInternal.h"
 #include "clang/ASTMatchers/Dynamic/VariantValue.h"
 
 #include <string>
 #include <algorithm>
 
+using namespace clang::ast_matchers;
 using namespace clang::ast_matchers::dynamic;
+
+namespace clang {
+namespace rewrite_tool {
 
 enum class MatcherType {
   fakeNode,
@@ -26,6 +25,7 @@ enum class MatcherType {
   cxxConstructExpr,
   cxxDefaultArgExpr,
   cxxFunctionalCastExpr,
+  cxxOperatorCallExpr,
   declRefExpr,
   equals,
   functionDecl,
@@ -318,6 +318,46 @@ public:
   }
 };
 
+VariantMatcher constructMatcher(StringRef MatcherName, int tab,
+                                Diagnostics *Error = nullptr);
+VariantMatcher constructMatcher(StringRef MatcherName,
+                                const VariantValue &Arg1,
+                                int tab,
+                                Diagnostics *Error = nullptr);
+VariantMatcher constructMatcher(StringRef MatcherName,
+                                const VariantValue &Arg1,
+                                const VariantValue &Arg2,
+                                int tab,
+                                Diagnostics *Error = nullptr);
+VariantMatcher constructMatcher(StringRef MatcherName,
+                                std::vector<VariantValue> args,
+                                int tab,
+                                Diagnostics* Error = nullptr);
+
+VariantMatcher constructBoundMatcher(StringRef MatcherName,
+                                     StringRef BoundName,
+                                     int tab,
+                                     Diagnostics *Error = nullptr);
+VariantMatcher constructBoundMatcher(StringRef MatcherName,
+                                     StringRef BoundName,
+                                     const VariantValue &Arg1,
+                                     int tab,
+                                     Diagnostics *Error = nullptr);
+VariantMatcher constructBoundMatcher(StringRef MatcherName,
+                                     StringRef BoundName,
+                                     const VariantValue &Arg1,
+                                     const VariantValue &Arg2,
+                                     int tab,
+                                     Diagnostics *Error = nullptr);
+VariantMatcher constructBoundMatcher(StringRef MatcherName,
+                                     StringRef BoundName,
+                                     std::vector<VariantValue> args,
+                                     int tab,
+                                     Diagnostics* Error = nullptr);
+
 VariantMatcher make_matcher(Node* root, int level);
+
+}
+} //namespace
 
 #endif // CLANG_CONSTRUCT_MATCHERS_H
