### 2022-07-27 ajohnson-uoregon/clang-rewrite-only#2 @ajohnson-uoregon
Updated to NCSA license: 225a318956e3586d and [c92ee871](https://github.com/ajohnson-uoregon/llvm-project/commit/ca5fed1db5c23c031fcbb6c8d8d26ab7c2da8253).
