### 2022-07-27 ajohnson-uoregon/clang-rewrite-only#2 @wimvanderbauwhede
[openjournals/joss-reviews#4590](https://github.com/openjournals/joss-reviews/issues/4590)

@ajohnson-uoregon Thanks, we can finalise the review now
