### 2022-07-27 70e1898
fixing formatting in readme and adding contribution guidelines
