### 2022-08-08 #4590 @danielskatz
Hi @wimvanderbauwhede & @tokuso2 - I've been on vacation the last weeks, but am now back and catching up.  It looks like you both are satisfied with this to proceed to publication based on your checklists.  Is this correct?
