### 2022-08-08 #4590 @danielskatz
@ajohnson-uoregon - in your paper.md, can you add the following line in the header after the bibliography line

```europar-doi: 10.1007/978-3-031-12597-3_12```

where the DOI should be the correct DOI for your Euro-Par paper.
