### 2022-08-11 #4590 @ajohnson-uoregon
I'm attempting to make a new archive on Zenodo, but running into errors with the file upload (due to an unstable internet connection, I think). I'll try again overnight when my computer isn't running a bunch of other stuff and we'll see how it goes.
