### 2022-08-12 #4590 @ajohnson-uoregon
Okay I gave up on Zenodo, the internet connection between them and me is too flaky. It's now archived on figshare: https://doi.org/10.6084/m9.figshare.20483967
